enable simple method to obtain configs
The config.xml is required to create new jobs, and can be the base of
automating job creation with proper standardization.

This commit saves a dummy config that can create a multibranch pipeline
job for this repo from the API.

It also adds a simple script that allows a user to build out and test
Jenkins jobs within docker, and copy the configuration over to a local
job-configs directory for whatever the desire may be. Such as utilizing
the config as a template for job creation automation.

Usage:

```
./bin/get-job-config source-job-name [target-filename]
```